fix: Use localStorage for memqueue tests instead of Redis for #162
